+++ trajectory_generator/include/trajectory_generator/trajectory_generator_diff_drive.h
@@ -12,15 +12,15 @@ class TrajectoryGeneratorDiffDrive
 public:
     TrajectoryGeneratorDiffDrive(void);
 
-    void set_optimization_param(const double, const double, const double);
-    void set_motion_param(const double, const double, const double, const double);
-    double generate_optimized_trajectory(const Eigen::Vector3d&, const MotionModelDiffDrive::ControlParams&, const double, const double, const int, MotionModelDiffDrive::ControlParams&, MotionModelDiffDrive::Trajectory&);
-    void get_jacobian(const double, const MotionModelDiffDrive::ControlParams&, const Eigen::Vector3d&, Eigen::Matrix3d&);
+    void set_optimization_param(const float, const float, const float);
+    void set_motion_param(const float, const float, const float, const float);
+    float generate_optimized_trajectory(const Eigen::Vector3f&, const MotionModelDiffDrive::ControlParams&, const float, const float, const int, MotionModelDiffDrive::ControlParams&, MotionModelDiffDrive::Trajectory&);
+    void get_jacobian(const float, const MotionModelDiffDrive::ControlParams&, const Eigen::Vector3f&, Eigen::Matrix3f&);
 
 private:
     MotionModelDiffDrive model;
 
-    Eigen::Vector3d h;
+    Eigen::Vector3f h;
 };
 
 #endif //__TRAJECTORY_GENERATOR_DIFF_DRIVE_H
