+++ trajectory_generator/src/trajectory_generator_diff_drive.cpp
@@ -6,28 +6,28 @@ TrajectoryGeneratorDiffDrive::TrajectoryGeneratorDiffDrive(void)
     h << 0.005, 0.005, 0.1;
 }
 
-void TrajectoryGeneratorDiffDrive::set_optimization_param(const double dkm, const double dkf, const double dsf)
+void TrajectoryGeneratorDiffDrive::set_optimization_param(const float dkm, const float dkf, const float dsf)
 {
     h << dkm, dkf, dsf;
 }
 
-void TrajectoryGeneratorDiffDrive::set_motion_param(const double max_yawrate, const double max_curvature, const double max_d_curvature, const double max_acceleration)
+void TrajectoryGeneratorDiffDrive::set_motion_param(const float max_yawrate, const float max_curvature, const float max_d_curvature, const float max_acceleration)
 {
     model.set_param(max_yawrate, max_curvature, max_d_curvature, max_acceleration);
 }
 
-double TrajectoryGeneratorDiffDrive::generate_optimized_trajectory(const Eigen::Vector3d& goal, const MotionModelDiffDrive::ControlParams& init_control_param, const double dt, const double tolerance, const int max_iteration, MotionModelDiffDrive::ControlParams& output, MotionModelDiffDrive::Trajectory& trajectory)
+float TrajectoryGeneratorDiffDrive::generate_optimized_trajectory(const Eigen::Vector3f& goal, const MotionModelDiffDrive::ControlParams& init_control_param, const float dt, const float tolerance, const int max_iteration, MotionModelDiffDrive::ControlParams& output, MotionModelDiffDrive::Trajectory& trajectory)
 {
-    Eigen::Vector3d cost(1e2, 1e2, 1e2);
-    double last_cost = cost.norm();
+    Eigen::Vector3f cost(1e2, 1e2, 1e2);
+    float last_cost = cost.norm();
 
-    double distance_to_goal = goal.segment(0, 2).norm();
+    float distance_to_goal = goal.segment(0, 2).norm();
 
     output = init_control_param;
 
     int count = 0;
 
-    Eigen::Matrix3d jacobian;
+    Eigen::Matrix3f jacobian;
 
     while(1){
         //double start = ros::Time::now().toSec();
@@ -41,7 +41,7 @@ double TrajectoryGeneratorDiffDrive::generate_optimized_trajectory(const Eigen::
         trajectory.trajectory.clear();
         trajectory.velocities.clear();
         trajectory.angular_velocities.clear();
-        double time = distance_to_goal / output.vel.v0;
+        float time = distance_to_goal / output.vel.v0;
         //std::cout << "bfr traj gen time: " << ros::Time::now().toSec() - start << "[s]" << std::endl;
         model.generate_trajectory(dt, output, trajectory);
         /*
@@ -58,8 +58,8 @@ double TrajectoryGeneratorDiffDrive::generate_optimized_trajectory(const Eigen::
         //std::cout << "j: \n" << jacobian << std::endl;
         //std::cout << "j^-1: \n" << jacobian.inverse() << std::endl;
         cost = goal - trajectory.trajectory.back();
-        Eigen::Vector3d dp = jacobian.inverse() * cost;
-        //Eigen::Vector3d dp = jacobian.lu().solve(cost);
+        Eigen::Vector3f dp = jacobian.inverse() * cost;
+        //Eigen::Vector3f dp = jacobian.lu().solve(cost);
         //std::cout << "jacobian inverse time: " << ros::Time::now().toSec() - start << "[s]" << std::endl;
         //std::cout << "cost: \n" << cost << std::endl;
         //std::cout << "dp: \n" << dp << std::endl;
@@ -77,7 +77,7 @@ double TrajectoryGeneratorDiffDrive::generate_optimized_trajectory(const Eigen::
             std::cout << "optimization error!!!" << std::endl;
             return -1;
         }
-        //std::cout << output.curv.km << ", " << output.curv.kf << ", " << output.curv.sf << std::endl; 
+        //std::cout << output.curv.km << ", " << output.curv.kf << ", " << output.curv.sf << std::endl;
 
         //std::cout << "count: " << count << std::endl;
         count++;
@@ -87,7 +87,7 @@ double TrajectoryGeneratorDiffDrive::generate_optimized_trajectory(const Eigen::
     return cost.norm();
 }
 
-void TrajectoryGeneratorDiffDrive::get_jacobian(const double dt, const MotionModelDiffDrive::ControlParams& control, const Eigen::Vector3d& h, Eigen::Matrix3d& j)
+void TrajectoryGeneratorDiffDrive::get_jacobian(const float dt, const MotionModelDiffDrive::ControlParams& control, const Eigen::Vector3f& h, Eigen::Matrix3f& j)
 {
     /*
      * h: (dkm, dkf, dsf)
@@ -95,26 +95,26 @@ void TrajectoryGeneratorDiffDrive::get_jacobian(const double dt, const MotionMod
     //std::cout << "j start" << std::endl;
     double start = ros::Time::now().toSec();
     MotionModelDiffDrive::CurvatureParams curv = control.curv;
-    Eigen::Vector3d x0;
+    Eigen::Vector3f x0;
     model.generate_last_state(dt, curv.sf, control.vel, curv.k0, curv.km - h(0), curv.kf, x0);
-    Eigen::Vector3d x1;
+    Eigen::Vector3f x1;
     model.generate_last_state(dt, curv.sf, control.vel, curv.k0, curv.km + h(0), curv.kf, x1);
 
-    Eigen::Vector3d dx_dkm;
+    Eigen::Vector3f dx_dkm;
     dx_dkm << (x1 - x0) / (2.0 * h(0));
     //std::cout << "dx_dkm time: " << ros::Time::now().toSec() - start << "[s]" << std::endl;
 
     model.generate_last_state(dt, curv.sf, control.vel, curv.k0, curv.km, curv.kf - h(1), x0);
     model.generate_last_state(dt, curv.sf, control.vel, curv.k0, curv.km, curv.kf + h(1), x1);
 
-    Eigen::Vector3d dx_dkf;
+    Eigen::Vector3f dx_dkf;
     dx_dkf << (x1 - x0) / (2.0 * h(1));
     //std::cout << "dx_dkf time: " << ros::Time::now().toSec() - start << "[s]" << std::endl;
 
     model.generate_last_state(dt, curv.sf - h(2), control.vel, curv.k0, curv.km, curv.kf, x0);
     model.generate_last_state(dt, curv.sf + h(2), control.vel, curv.k0, curv.km, curv.kf, x1);
 
-    Eigen::Vector3d dx_dsf;
+    Eigen::Vector3f dx_dsf;
     dx_dsf << (x1 - x0) / (2.0 * h(2));
     //std::cout << "dx_dsf time: " << ros::Time::now().toSec() - start << "[s]" << std::endl;
